Ensure tail indices are not lost
Previously, read_ib would only keep indices that were part of a triangle - if the total was not divisible by 3, the last 1-2 would be lost.  Now the final indices will be added to a list which may be shorter than 3 elements, so that formats other than trianglelist can be read (although grouping will be irrelevant and all the lists will need to be chained back together).