Fixed crash with multi-input slot VBs
Submesh "vertexBufferIndex" actually points to vertex attributes, not vertex buffer.  Vertex attributes has a list that points to vertex buffer.  The script will automatically merge the input slots into a single buffer, since the Blender plugin does not support multiple input slots anyway.  When reconstructed, it will adhere to the newer format of one input slot, but this should not be an issue in-game (I hope).