Support old-style Oid, remove trianglestrip option
Trianglestrip output is still possible at the function call level, since GLTF interpreter will still take the triangle strip (and advanced users may still want to enable the output).  But since it's not feasible to repack trianglestrip or to view it in Blender, will leave it out of the commandline options to prevent errors.