Add global variable for vertex color behavior
The default in the past has been to discard vertex color, since the KT engine does not actually use it for vertex color.  However, for specific modding operations, it is important to preserve all semantics in the model, so this makes it easier to set the default behavior at a user level.