Deal with small repeating vertex buffers
An example g1m (0x7C45686C) from DWO has a vertex buffer, one with position/normal/etc for all the vertices in the first input slot , and the second input slot with a single set of data meant to be repeated over and over for all the vertices.  This compensates for that.  I do not know if modulus is the proper way to deal with this because I only have one data point, I can change this later as needed.